requirements.txt
gunicorn==21.2.0
absl-py==2.1.0
aiofiles==22.1.0
aiosqlite==0.19.0
altgraph==0.17.3
anyio==3.7.0
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
arrow==1.2.3
astunparse==1.6.3
attrs==23.1.0
Babel==2.12.1
backcall==0.2.0
beautifulsoup4==4.12.2
bleach==6.0.0
branca==0.7.1
CacheControl==0.12.11
cached-property==1.5.2
cachetools==5.3.0
certifi==2022.12.7
cffi==1.15.1
chardet==3.0.4
charset-normalizer==3.1.0
click==8.1.3
cmake==3.25.0
colorama==0.4.6
cryptography==41.0.7
cvzone==1.5.6
cycler==0.11.0
debugpy==1.6.7
decorator==5.1.1
defusedxml==0.7.1
dlib==19.24.0
dronekit==2.9.2
entrypoints==0.4
et-xmlfile==1.1.0
exceptiongroup==1.1.1
face-recognition==1.3.0
face-recognition-models==0.3.0
fastjsonschema==2.17.1
firebase-admin==6.1.0
Flask==2.2.5
flatbuffers==24.3.25
folium==0.16.0
fonttools==4.38.0
fqdn==1.5.1
future==0.18.3
gast==0.4.0
geographiclib==2.0
geopy==2.4.1
google-api-core==2.11.0
google-api-python-client==2.84.0
google-auth==2.17.2
google-auth-httplib2==0.1.0
google-auth-oauthlib==0.4.6
google-cloud-core==2.3.2
google-cloud-firestore==2.11.0
google-cloud-storage==2.8.0
google-crc32c==1.5.0
google-pasta==0.2.0
google-resumable-media==2.4.1
googleapis-common-protos==1.59.0
googletrans==4.0.0rc1
grpcio==1.53.0
grpcio-status==1.53.0
h11==0.9.0
h2==3.2.0
h5py==3.8.0
hpack==3.0.0
hstspreload==2024.1.5
httpcore==0.9.1
httplib2==0.22.0
httpx==0.13.3
hyperframe==5.2.0
idna==2.10
importlib-metadata==6.0.0
importlib-resources==5.12.0
install==1.3.5
ipykernel==6.16.2
ipython==7.34.0
ipython-genutils==0.2.0
isoduration==20.11.0
itsdangerous==2.1.2
jedi==0.18.2
Jinja2==3.1.2
joblib==1.2.0
json5==0.9.14
jsonpointer==2.3
jsonschema==4.17.3
jupyter-events==0.6.3
jupyter-server==1.24.0
jupyter-ydoc==0.2.4
jupyter_client==7.4.9
jupyter_core==4.12.0
jupyter_server_fileid==0.9.0
jupyter_server_ydoc==0.8.0
jupyterlab==3.6.4
jupyterlab-pygments==0.2.2
jupyterlab_server==2.23.0
keras==2.11.0
kiwisolver==1.4.4
libclang==18.1.1
lxml==4.9.2
Markdown==3.4.4
MarkupSafe==2.1.3
matplotlib==3.5.3
matplotlib-inline==0.1.6
mistune==2.0.5
monotonic==1.6
msgpack==1.0.5
mysql-connector==2.2.9
mysql-connector-python==8.0.33
nbclassic==1.0.0
nbclient==0.7.4
nbconvert==7.5.0
nbformat==5.8.0
nest-asyncio==1.5.6
nltk==3.8.1
notebook==6.5.4
notebook_shim==0.2.3
numpy==1.21.6
oauthlib==3.2.2
opencv-python==4.2.0.32
openpyxl==3.1.2
opt-einsum==3.3.0
outcome==1.3.0.post0
packaging==23.1
pandas==1.1.5
pandocfilters==1.5.0
parso==0.8.3
pefile==2023.2.7
php==1.2.1
pickleshare==0.7.5
Pillow==9.4.0
pkgutil_resolve_name==1.3.10
prometheus-client==0.17.0
prompt-toolkit==3.0.38
proto-plus==1.22.2
protobuf==3.19.6
psutil==5.9.5
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
pygame==2.5.2
Pygments==2.15.1
pyinstaller==5.13.0
pyinstaller-hooks-contrib==2023.6
PyJWT==2.6.0
pymavlink==2.4.37
pyOpenSSL==23.3.0
pyparsing==3.0.9
pyrsistent==0.19.3
python-dateutil==2.8.2
python-json-logger==2.0.7
pytz==2023.3
pywin32==306
pywin32-ctypes==0.2.2
pywinpty==2.0.10
PyYAML==6.0
pyzmq==25.1.0
regex==2023.10.3
requests==2.28.2
requests-oauthlib==2.0.0
rfc3339-validator==0.1.4
rfc3986==1.5.0
rfc3986-validator==0.1.1
rsa==4.9
scikit-learn==1.0.2
scipy==1.7.3
selenium==4.0.0b4
Send2Trash==1.8.2
six==1.16.0
sniffio==1.3.0
sortedcontainers==2.4.0
soupsieve==2.4.1
tensorboard==2.11.2
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tensorflow==2.11.0
tensorflow-estimator==2.11.0
tensorflow-intel==2.11.0
tensorflow-io-gcs-filesystem==0.31.0
termcolor==2.3.0
terminado==0.17.1
textblob==0.17.1
threadpoolctl==3.1.0
tinycss2==1.2.1
tk==0.1.0
tomli==2.0.1
tornado==6.2
tqdm==4.66.1
traitlets==5.9.0
trio==0.22.2
trio-websocket==0.11.1
typing_extensions==4.4.0
uri-template==1.2.0
uritemplate==4.1.1
urllib3==1.26.15
urllib3-secure-extra==0.1.0
wcwidth==0.2.6
webcolors==1.13
webencodings==0.5.1
websocket-client==1.5.3
Werkzeug==2.2.3
wrapt==1.16.0
wsproto==1.2.0
xlrd==2.0.1
xyzservices==2023.2.0
y-py==0.5.9
ypy-websocket==0.8.2
zipp==3.13.0
